% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtools.R
\name{get_path}
\alias{get_path}
\title{Tools for directory management}
\usage{
get_path()
}
\description{
Tools for directory management
}
\keyword{active}
\keyword{directory}
\keyword{document}
\keyword{get}
\keyword{path}
\keyword{working}
